PYTHON-6015 Consolidate topology module-level helpers into shared class
